Isolate netty package in one file
